requirements.txt
acres==0.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
asyncpg==0.30.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
exceptiongroup==1.3.1
fastapi==0.121.3
filelock==3.20.0
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.52.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
isodate==0.7.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.80
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.11
langsmith==0.4.46
looseversion==1.3.0
lxml==6.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psycopg2-binary==2.9.11
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pydot==4.0.1
PyMuPDF==1.26.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.4.0
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.15.3
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
tenacity==9.1.2
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
zstandard==0.25.0
